Some other functions have been extracted from act.item.cpp into separate files
