+++ CMakeLists.txt
@@ -197,7 +197,7 @@ set(SOURCES
 		src/game_skills/block.cpp
 		src/game_skills/parry.cpp
 		src/game_skills/pick.cpp
-		src/game_skills/poison.cpp
+		src/game_mechanics/poison.cpp
 		src/game_skills/protect.cpp
 		src/game_skills/expendientcut.cpp
 		src/game_skills/turnundead.cpp
@@ -253,7 +253,7 @@ set(SOURCES
 		src/cmd/give.cpp
 		src/cmd/trample.cpp
 		src/game_skills/campfire.cpp
-		src/game_skills/firstaid.cpp)
+		src/game_skills/firstaid.cpp src/game_skills/sharpening.cpp src/game_skills/armoring.cpp src/cmd/get.cpp src/cmd/equip.cpp src/cmd/put.cpp src/cmd/remove.cpp)
 
 set(HEADERS
 		src/cmd/flee.h
@@ -454,7 +454,7 @@ set(HEADERS
 		src/game_skills/backstab.h
 		src/game_skills/bash.h
 		src/game_skills/stun.h
-		src/game_skills/poison.h
+		src/game_mechanics/poison.h
 		src/game_skills/resque.h
 		src/game_skills/kick.h
 		src/game_skills/strangle.h
@@ -486,7 +486,7 @@ set(HEADERS
 		src/cmd/give.h
 		src/cmd/trample.h
 		src/game_skills/campfire.h
-		src/game_skills/firstaid.h)
+		src/game_skills/firstaid.h src/game_skills/sharpening.h src/game_skills/armoring.h src/cmd/get.h src/cmd/equip.h src/cmd/put.h src/cmd/remove.h)
 
 set(CONFIGURATION_FILES lib.template/misc/configuration.xml)
 # Build types
