+++ src/cmd/put.h
@@ -1,0 +1,10 @@
+#ifndef BYLINS_SRC_CMD_PUT_H_
+#define BYLINS_SRC_CMD_PUT_H_
+
+class CharData;
+
+void do_put(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_SRC_CMD_PUT_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
