+++ src/game_skills/armoring.h
@@ -1,0 +1,10 @@
+#ifndef BYLINS_SRC_GAME_SKILLS_ARMORING_H_
+#define BYLINS_SRC_GAME_SKILLS_ARMORING_H_
+
+class CharData;
+
+void DoArmoring(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_SRC_GAME_SKILLS_ARMORING_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
