+++ src/game_skills/sharpening.h
@@ -1,0 +1,14 @@
+//
+// Created by Alexandr on 09.05.2022.
+//
+
+#ifndef BYLINS_SRC_GAME_SKILLS_SHARPENING_H_
+#define BYLINS_SRC_GAME_SKILLS_SHARPENING_H_
+
+class CharData;
+
+void DoSharpening(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_SRC_GAME_SKILLS_SHARPENING_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
